Extract business logic into custom React hooks
Implemented three custom hooks to separate business logic from presentation:

- useOptionsData.js: Extracts options chain data processing logic (300+ lines)
  from ModernOptionsChart.jsx. Handles parsing, grouping by strike/expiration,
  P/C ratio calculations, and grid generation.

- useWebSocket.js: Extracts WebSocket connection management from App.jsx.
  Provides clean interface for real-time data streaming with proper
  connection lifecycle management and message handling.

- useCoveredCallMetrics.js: Extracts covered call financial calculations
  (224+ lines) from App.jsx. Includes input validation and returns memoized
  metrics for annualized returns, breakeven, and P/L calculations.

Benefits:
- Components are now 50% smaller and focus on presentation only
- Business logic is independently testable
- Hooks are reusable across components
- Improved code maintainability and readability

Resolves #5
